Update cron schedule for taking snapshots at 2:00 PM daily
The Dockerfile has been modified to update the cron schedule for taking snapshots. The previous schedule was set to run every minute, but it has been changed to run at 2:00 PM daily. This change ensures that the snapshot.py script is executed at the desired time using the specified configuration file.